+++ src/applications/vaos/appointment-list/components/ConfirmedAppointmentDetailsPage/VideoVisitProvider.jsx
@@ -1,10 +1,18 @@
 import React from 'react';
 
-export default function VideoVisitProvider({ providers, isPast }) {
+export default function VideoVisitProvider({ appointment }) {
+  const providers = appointment.videoData.providers;
+
+  if (!providers?.length) {
+    return null;
+  }
+
   return (
-    <>
+    <div className="vads-u-margin-top--2">
       <h2 className="vads-u-font-size--base vads-u-font-family--sans vads-u-margin-bottom--0">
-        {!isPast ? 'You’ll be meeting with' : 'You met with'}
+        {!appointment.vaos.isPastAppointment
+          ? 'You’ll be meeting with'
+          : 'You met with'}
       </h2>
       {/* eslint-disable-next-line jsx-a11y/no-redundant-roles */}
       <ul className="usa-unstyled-list" role="list">
@@ -12,6 +20,6 @@ export default function VideoVisitProvider({ providers, isPast }) {
           <li key={index}>{provider.display}</li>
         ))}
       </ul>
-    </>
+    </div>
   );
 }
